refactor!: module prefix `cli` rename file
each cli command file should be named `cli_name.py` for consistency
